+++ back/src/main/java/com/tdevilleduc/urthehero/back/service/impl/PersonService.java
@@ -3,7 +3,6 @@ package com.tdevilleduc.urthehero.back.service.impl;
 import com.tdevilleduc.urthehero.back.convertor.PersonConvertor;
 import com.tdevilleduc.urthehero.back.dao.PersonDao;
 import com.tdevilleduc.urthehero.back.exceptions.PersonNotFoundException;
-import com.tdevilleduc.urthehero.back.model.Page;
 import com.tdevilleduc.urthehero.back.model.Person;
 import com.tdevilleduc.urthehero.back.model.PersonDTO;
 import com.tdevilleduc.urthehero.back.service.IPersonService;
@@ -19,6 +18,8 @@ import java.util.List;
 import java.util.Optional;
 
 import static com.tdevilleduc.urthehero.back.config.ResilienceConstants.INSTANCE_PERSON_SERVICE;
+import static com.tdevilleduc.urthehero.back.constant.ApplicationConstants.CHECK_PERSONID_PARAMETER_MANDATORY;
+import static com.tdevilleduc.urthehero.back.constant.ApplicationConstants.ERROR_MESSAGE_PERSON_DOESNOT_EXIST;
 
 @Slf4j
 @Service
@@ -34,7 +35,7 @@ public class PersonService implements IPersonService {
     public boolean exists(final Integer personId) {
         Optional<Person> person = personDao.findById(personId);
         if (person.isEmpty()) {
-            log.error("La personne avec l'id {} n'existe pas", personId);
+            log.error(ERROR_MESSAGE_PERSON_DOESNOT_EXIST, personId);
             return false;
         }
         return true;
@@ -52,7 +53,7 @@ public class PersonService implements IPersonService {
 
     @CircuitBreaker(name = INSTANCE_PERSON_SERVICE, fallbackMethod = "emptyPerson")
     public Optional<Person> findById(final Integer personId) {
-        Assert.notNull(personId, "The personId parameter is mandatory !");
+        Assert.notNull(personId, CHECK_PERSONID_PARAMETER_MANDATORY);
         return personDao.findById(personId);
     }
 
@@ -79,12 +80,12 @@ public class PersonService implements IPersonService {
     }
 
     public void delete(Integer personId) {
-        Assert.notNull(personId, "The personId parameter is mandatory !");
+        Assert.notNull(personId, CHECK_PERSONID_PARAMETER_MANDATORY);
         Optional<Person> optional = findById(personId);
         optional
             .ifPresentOrElse(person -> personDao.delete(person),
                 () -> {
-                    throw new PersonNotFoundException(MessageFormatter.format("La personne avec l'id {} n'existe pas", personId).getMessage());
+                    throw new PersonNotFoundException(MessageFormatter.format(ERROR_MESSAGE_PERSON_DOESNOT_EXIST, personId).getMessage());
                 }
         );
     }